refactor: Move permission router inside routers folder, add cognify router
Moved permissions router in it's routers folder, added cognify router

Refactor #COG-334
