+++ cognee/api/v1/permissions/routers/__init__.py
@@ -1,0 +1,1 @@
+from .get_permissions_router import get_permissions_router
